+++ src/structures/DBBase.ts
@@ -1,0 +1,7 @@
+class Model<T> {}
+
+export = class DB {
+  public collections: Map<string, Model<any>> = new Map()
+
+  constructor() {}
+}