+++ src/structures/DBProvider.d.ts
@@ -1,0 +1,28 @@
+/// <reference types="node" />
+import mongoose from 'mongoose';
+import { EventEmitter } from 'events';
+declare const _default: {
+    new (): {
+        connection: mongoose.Connection;
+        _db?: any;
+        db: any;
+        addListener(event: string | symbol, listener: (...args: any[]) => void): any;
+        on(event: string | symbol, listener: (...args: any[]) => void): any;
+        once(event: string | symbol, listener: (...args: any[]) => void): any;
+        removeListener(event: string | symbol, listener: (...args: any[]) => void): any;
+        off(event: string | symbol, listener: (...args: any[]) => void): any;
+        removeAllListeners(event?: string | symbol | undefined): any;
+        setMaxListeners(n: number): any;
+        getMaxListeners(): number;
+        listeners(event: string | symbol): Function[];
+        rawListeners(event: string | symbol): Function[];
+        emit(event: string | symbol, ...args: any[]): boolean;
+        listenerCount(type: string | symbol): number;
+        prependListener(event: string | symbol, listener: (...args: any[]) => void): any;
+        prependOnceListener(event: string | symbol, listener: (...args: any[]) => void): any;
+        eventNames(): (string | symbol)[];
+    };
+    listenerCount(emitter: EventEmitter, event: string | symbol): number;
+    defaultMaxListeners: number;
+};
+export = _default;